Refactor(validator): Separate ValidationResult data class from InputParser.kt into ValidateResult.kt
