+++ src/main/kotlin/lotto/view/InputParser.kt
@@ -1,25 +1,8 @@
 package lotto.view
 
 import lotto.error.*
+import lotto.validator.ValidationResult
 import lotto.validator.Validator
-
-
-data class ValidationResult<T>(
-	val errorType: ErrorType,
-	val value: T?
-) {
-	fun isValid(): Boolean {
-		if (errorType.isStatusFailure()) return false
-		if (errorType.isStatusSuccess() && value != null) return true
-		ExceptionHandler.throwIf(LogicError.INVALID_LOGIC)
-		return false
-	}
-
-	fun toNullFree(): T {
-		return value ?: throw UnexpectedException(LogicError.CONVERSION_FAILED.toMessage())
-	}
-
-}
 
 object InputParser {
 
@@ -27,13 +10,14 @@ object InputParser {
 		val validationResult = fromStringToAmount(input)
 		if (validationResult.isValid()) {
 			val validatedAmount = validationResult.toNullFree()
-			val validateStatus = Validator.checkAmount(validatedAmount)
-			if (validateStatus.isStatusSuccess()) {
+			val error = Validator.checkAmount(validatedAmount)
+			if (error.isStatusSuccess()) {
 				return validatedAmount
 			}
-			ExceptionHandler.throwIf(validateStatus)
+			ExceptionHandler.throwIf(error)
 		}
-		throw UnexpectedException(LogicError.DEFAULT.toMessage())
+		ExceptionHandler.throwIf(validationResult.errorType)
+		throw UnreachableCodeException("This code should not be reachable")
 	}
 
 	private fun fromStringToAmount(input: String): ValidationResult<Long> {
@@ -51,21 +35,22 @@ object InputParser {
 		val validationResult = fromStringToWinning(input)
 		if (validationResult.isValid()) {
 			val list = validationResult.toNullFree()
-			val validateStatus = Validator.checkWinningNumbers(list)
-			if (validateStatus.isStatusSuccess()) {
+			val error = Validator.checkWinningNumbers(list)
+			if (error.isStatusSuccess()) {
 				return list
 			}
-			ExceptionHandler.throwIf(validateStatus)
+			ExceptionHandler.throwIf(error)
 		}
-		throw UnexpectedException(LogicError.DEFAULT.toMessage())
+		ExceptionHandler.throwIf(validationResult.errorType)
+		throw UnreachableCodeException("This code should not be reachable")
 	}
 
 	// TODO: refactor: to small functions
 	private fun fromStringToWinning(input: String): ValidationResult<List<Int>> {
 		val winningStrings = parseCommaSeparatedStrings(input)
-		val digitStatus = Validator.checkAllDigit(winningStrings)
-		if (digitStatus.isStatusFailure())
-			return ValidationResult(digitStatus, null)
+		val error = Validator.checkNumeric(winningStrings)
+		if (error.isStatusFailure())
+			return ValidationResult(error, null)
 		val listOrNull = winningStrings.map { it.toIntOrNull() }
 		if (listOrNull.any { it == null }) {
 			return ValidationResult(ParseError.INVALID_RANGE, null)
@@ -80,7 +65,7 @@ object InputParser {
 	private fun parseCommaSeparatedStrings(input: String): List<String> {
 		val parsed = input.split(',')
 			.map { it.trim() }
-		return (parsed)
+		return parsed
 	}
 }
